+++ biosimulations/libs/datamodel/core/src/lib/common/chart-data-field.dto.ts
@@ -1,5 +1,5 @@
-import { ChartDataFieldShape } from '../enums/chart-type-data-field-shape';
-import { ChartDataFieldType } from '../enums/chart-type-data-field-type';
+import { ChartDataFieldShape } from '../enums/chart-data-field-shape';
+import { ChartDataFieldType } from '../enums/chart-data-field-type';
 import { DTO } from '@biosimulations/datamodel/utils';
 
 export interface ChartDataFieldCore {
