+++ biosimulations/libs/datamodel/core/src/lib/enums/chart-data-field-type.ts
