+++ biosimulations/libs/datamodel/core/src/lib/enums/index.ts
@@ -5,8 +5,8 @@ export {
 } from './project-product-type';
 
 export { AccessLevel, accessLevels } from './access-level';
-export { ChartDataFieldShape } from './chart-type-data-field-shape';
-export { ChartDataFieldType } from './chart-type-data-field-type';
+export { ChartDataFieldShape } from './chart-data-field-shape';
+export { ChartDataFieldType } from './chart-data-field-type';
 export { ChartTypeType } from './chart-type-type';
 export { License } from './licence';
 export { ResourceType } from './resource-type';
